fix: bind pending port release to script generation
